+++ include/linux/vt_kern.h
@@ -38,7 +38,7 @@ int con_get_cmap(unsigned char __user *cmap);
 void scrollback(struct vc_data *vc);
 void scrollfront(struct vc_data *vc, int lines);
 void clear_buffer_attributes(struct vc_data *vc);
-void update_region(struct vc_data *vc, unsigned long start, int count);
+void update_region(struct vc_data *vc, uintptr_t start, int count);
 void redraw_screen(struct vc_data *vc, int is_switch);
 #define update_screen(x) redraw_screen(x, 0)
 #define switch_screen(x) redraw_screen(x, 1)
